Create custom hooks to get all triggers enabled values
